Crate concordium_rust_sdk

A library for interacting with the Concordium blockchain.

To use this in your project the recommended setup is to clone the repository (e.g., add it as a submodule), and then declare a local dependency in your Cargo.toml, e.g.,

[dependencies]
concordium-rust-sdk = { path = "./deps/concordium-rust-sdk", version = "1" }

The library is structured around multiple modules.

endpoints contains the main entrypoint to the library. In particular it contains the Client struct which maintains a connection to the node, and supports queries and node manipulation.

constants contains a number of constants and type definitions that are relevant when using the chain. In particular DEFAULT_NETWORK_ID.

types contains most type definitions to model responses as well as types defining transactions. The latter are in a submodule types::transactions.

In addition to these, the library re-exports a number of core crates that implement the core cryptographic protocols of the Concordium blockchain.

id is the implementation of most of the protocols in the identity layer

common has some common type definitions, as well as traits and helpers for binary serialization

encrypted_transfers implements structures and zero knowledge proofs related to encrypted transfers

eddsa_ed25519 is a re-export of the signature scheme used for blocks and accounts on the Concordium blockchain.

aggregate_sig is a re-export of the BLS signature scheme, used by the finalizers.

ecvrf is a re-export of the implementation of the VRF function used to determine lottery winners in consensus.

Re-exports

pub use id;

pub use crypto_common as common;

pub use encrypted_transfers;

pub use aggregate_sig;

pub use eddsa_ed25519;

pub use ecvrf;

Modules

cis2

Types and functions for working with CIS-2 smart contracts. This module contains types and functions for interacting with smart contracts following the CIS-2 specification.

constants

Various constants and types that apply to the chain.

endpoints

Wrapper for the node’s GRPC API. The return values are parsed and wrapped in structured values.

types

Type definitions used throughout the rest of the SDK.