Module concordium_rust_sdk::constants

Various constants and types that apply to the chain.

Constants

DEFAULT_NETWORK_ID
Identifier of the default network over which messages are transmitted. This is also the only currently supported network.

MAX_MEMO_SIZE
Max allowed memo size.

MAX_PARAMETER_LEN
Maximum allowed length of a smart contract parameter. This must be kept in sync with maxParameterLen.

MAX_PAYLOAD_SIZE
Maximum size of a transaction payload.

MAX_REGISTERED_DATA_SIZE
Maximum allowed size of data to register via the register data transaction.

MAX_URL_TEXT_LENGTH
The maximum allowed length of a UrlText in bytes.

MAX_WASM_MODULE_SIZE
Maximum allowed size of the Wasm module to deploy on the chain.

MIN_NONCE
Minimum valid transaction nonce. Nonces must be strictly sequential starting with MIN_NONCE.

SHA256
Size of the sha256 digest in bytes.

Type Definitions

EncryptedAmountsCurve
Curve used for encrypted transfers. This is the same as the anonymity revoker curve.