Enum concordium_rust_sdk::types::RejectReason

pub enum RejectReason {

    ModuleNotWF,
    ModuleHashAlreadyExists {
        contents: ModuleRef,
    },
    InvalidAccountReference {
        contents: AccountAddress,
    },
    InvalidInitMethod {
        contents: (ModuleRef, InitName),
    },
    InvalidReceiveMethod {
        contents: (ModuleRef, ReceiveName),
    },
    InvalidModuleReference {
        contents: ModuleRef,
    },
    InvalidContractAddress {
        contents: ContractAddress,
    },
    RuntimeFailure,
    AmountTooLarge {
        contents: (Address, Amount),
    },
    SerializationFailure,
    OutOfEnergy,
    RejectedInit {
        reject_reason: i32,
    },
    RejectedReceive {
        reject_reason: i32,
        contract_address: ContractAddress,
        receive_name: ReceiveName,
        parameter: Parameter,
    },
    NonExistentRewardAccount {
        contents: AccountAddress,
    },
    InvalidProof,
    AlreadyABaker {
        contents: BakerId,
    },
    NotABaker {
        contents: AccountAddress,
    },
    InsufficientBalanceForBakerStake,
    StakeUnderMinimumThresholdForBaking,
    BakerInCooldown,
    DuplicateAggregationKey {
        contents: Box<BakerAggregationVerifyKey>,
    },
    NonExistentCredentialID,
    KeyIndexAlreadyInUse,
    InvalidAccountThreshold,
    InvalidCredentialKeySignThreshold,
    InvalidEncryptedAmountTransferProof,
    InvalidTransferToPublicProof,
    EncryptedAmountSelfTransfer {
        contents: AccountAddress,
    },
    InvalidIndexOnEncryptedTransfer,
    ZeroScheduledAmount,
    NonIncreasingSchedule,
    FirstScheduledReleaseExpired,
    ScheduledSelfTransfer {
        contents: AccountAddress,
    },
    InvalidCredentials,
    DuplicateCredIDs {
        contents: Vec<CredentialRegistrationID>,
    },
    NonExistentCredIDs {
        contents: Vec<CredentialRegistrationID>,
    },
    RemoveFirstCredential,
    CredentialHolderDidNotSign,
    NotAllowedMultipleCredentials,
    NotAllowedToReceiveEncrypted,
    NotAllowedToHandleEncrypted,
    MissingBakerAddParameters,
    FinalizationRewardCommissionNotInRange,
    BakingRewardCommissionNotInRange,
    TransactionFeeCommissionNotInRange,
    AlreadyADelegator,
    InsufficientBalanceForDelegationStake,
    MissingDelegationAddParameters,
    InsufficientDelegationStake,
    DelegatorInCooldown,
    NotADelegator {
        address: AccountAddress,
    },
    DelegationTargetNotABaker {
        target: BakerId,
    },
    StakeOverMaximumThresholdForPool,
    PoolWouldBecomeOverDelegated,
    PoolClosed,

}

A reason for why a transaction was rejected. Rejected means included in a block, but the desired action was not achieved. The only effect of a rejected transaction is payment.

NOTE: Some of the variant definitions can look peculiar, but they are made to be compatible with the serialization of the Haskell datatype.

Variants

ModuleNotWF

Error raised when validating the Wasm module.

ModuleHashAlreadyExists

As the name says.

Fields of ModuleHashAlreadyExists

contents: ModuleRef

InvalidAccountReference

Account does not exist.

Fields of InvalidAccountReference

contents: AccountAddress

InvalidInitMethod

Reference to a non-existing contract init method.

Fields of InvalidInitMethod

contents: (ModuleRef, InitName)

InvalidReceiveMethod

Reference to a non-existing contract receive method.

Fields of InvalidReceiveMethod

contents: (ModuleRef, ReceiveName)

InvalidModuleReference

Reference to a non-existing module.

Fields of InvalidModuleReference

contents: ModuleRef

InvalidContractAddress

Contract instance does not exist.

Fields of InvalidContractAddress

contents: ContractAddress

RuntimeFailure

Runtime exception occurred when running either the init or receive method.

AmountTooLarge

When one wishes to transfer an amount from A to B but there are not enough funds on account/contract A to make this possible. The data are the from address and the amount to transfer.

Fields of AmountTooLarge

contents: (Address, Amount)

SerializationFailure

Serialization of the body failed.

OutOfEnergy

We ran of out energy to process this transaction.

RejectedInit

Rejected due to contract logic in init function of a contract.

Fields of RejectedInit

reject_reason: i32

RejectedReceive

Fields of RejectedReceive

reject_reason: i32contract_address: ContractAddressreceive_name: ReceiveNameparameter: Parameter

NonExistentRewardAccount

Reward account desired by the baker does not exist.

Fields of NonExistentRewardAccount

contents: AccountAddress

InvalidProof

Proof that the baker owns relevant private keys is not valid.

AlreadyABaker

Tried to add baker for an account that already has a baker

Fields of AlreadyABaker

contents: BakerId

NotABaker

Tried to remove a baker for an account that has no baker

Fields of NotABaker

contents: AccountAddress

InsufficientBalanceForBakerStake

The amount on the account was insufficient to cover the proposed stake

StakeUnderMinimumThresholdForBaking

The amount provided is under the threshold required for becoming a baker

BakerInCooldown

The change could not be made because the baker is in cooldown for another change

DuplicateAggregationKey

A baker with the given aggregation key already exists

Fields of DuplicateAggregationKey

contents: Box<BakerAggregationVerifyKey>

NonExistentCredentialID

Encountered credential ID that does not exist

KeyIndexAlreadyInUse

Attempted to add an account key to a key index already in use

InvalidAccountThreshold

When the account threshold is updated, it must not exceed the amount of existing keys

InvalidCredentialKeySignThreshold

When the credential key threshold is updated, it must not exceed the amount of existing keys

InvalidEncryptedAmountTransferProof

Proof for an encrypted amount transfer did not validate.

InvalidTransferToPublicProof

Proof for a secret to public transfer did not validate.

EncryptedAmountSelfTransfer

Account tried to transfer an encrypted amount to itself, that’s not allowed.

Fields of EncryptedAmountSelfTransfer

contents: AccountAddress

InvalidIndexOnEncryptedTransfer

The provided index is below the start index or above startIndex + length incomingAmounts

ZeroScheduledAmount

The transfer with schedule is going to send 0 tokens

NonIncreasingSchedule

The transfer with schedule has a non strictly increasing schedule

FirstScheduledReleaseExpired

The first scheduled release in a transfer with schedule has already expired

ScheduledSelfTransfer

Account tried to transfer with schedule to itself, that’s not allowed.

Fields of ScheduledSelfTransfer

contents: AccountAddress

InvalidCredentials

At least one of the credentials was either malformed or its proof was incorrect.

DuplicateCredIDs

Some of the credential IDs already exist or are duplicated in the transaction.

Fields of DuplicateCredIDs

contents: Vec<CredentialRegistrationID>

NonExistentCredIDs

A credential id that was to be removed is not part of the account.

Fields of NonExistentCredIDs

contents: Vec<CredentialRegistrationID>

RemoveFirstCredential

Attemp to remove the first credential

CredentialHolderDidNotSign

The credential holder of the keys to be updated did not sign the transaction

NotAllowedMultipleCredentials

Account is not allowed to have multiple credentials because it contains a non-zero encrypted transfer.

NotAllowedToReceiveEncrypted

The account is not allowed to receive encrypted transfers because it has multiple credentials.

NotAllowedToHandleEncrypted

The account is not allowed to send encrypted transfers (or transfer from/to public to/from encrypted)

MissingBakerAddParameters

A configure baker transaction is missing one or more arguments in order to add a baker.

FinalizationRewardCommissionNotInRange

Finalization reward commission is not in the valid range for a baker

BakingRewardCommissionNotInRange

Baking reward commission is not in the valid range for a baker

TransactionFeeCommissionNotInRange

Transaction fee commission is not in the valid range for a baker

AlreadyADelegator

Tried to add baker for an account that already has a delegator.

InsufficientBalanceForDelegationStake

The amount on the account was insufficient to cover the proposed stake.

MissingDelegationAddParameters

A configure delegation transaction is missing one or more arguments in order to add a delegator.

InsufficientDelegationStake

Delegation stake when adding a delegator was 0.

DelegatorInCooldown

Account is not a delegation account.

NotADelegator

Account is not a delegation account.

Fields of NotADelegator

address: AccountAddress

DelegationTargetNotABaker

Delegation target is not a baker

Fields of DelegationTargetNotABaker

target: BakerId

StakeOverMaximumThresholdForPool

The amount would result in pool capital higher than the maximum threshold.

PoolWouldBecomeOverDelegated

The amount would result in pool with a too high fraction of delegated capital.

PoolClosed

The pool is not open to delegators.