Module concordium_rust_sdk::types::smart_contracts

Types related to smart contracts.

Re-exports

pub use concordium_contracts_common;

Structs

ContractContext
Data needed to invoke the contract.

ContractEvent
An event logged by a smart contract initialization.

InitName
FIXME: Add structure.

ModuleSource
Unparsed Wasm module source. FIXME: Make this structured based on what we have in wasm-chain-integration.

Parameter
A smart contract parameter. The Default implementation produces an empty parameter.

ReceiveName
FIXME: Add structure.

ReturnValue

WasmModule
Unparsed module with a version indicating what operations are allowed. FIXME: Make this structured based on what we have in wasm-chain-integration.

Enums

InstanceInfo
Information about an existing smart contract instance.

InvokeContractResult

WasmVersion

Constants

DEFAULT_INVOKE_ENERGY

Type Definitions

ModuleRef
Reference to a deployed Wasm module on the chain.