Trait concordium_rust_sdk::types::smart_contracts::concordium_contracts_common::Deserial

pub trait Deserial {
    fn deserial<R>(_source: &mut R) -> Result<Self, ParseError>
    where
        R: Read;
}

The Deserial trait provides a means of reading structures from byte-sources (Read).

Can be derived using #[derive(Deserial)] for most cases.

Required methods

fn deserial<R>(_source: &mut R) -> Result<Self, ParseError> where
R: Read,

Attempt to read a structure from a given source, failing if an error occurs during deserialization or reading.