Trait concordium_rust_sdk::types::smart_contracts::concordium_contracts_common::Serial

pub trait Serial {
    fn serial<W>(&self, _out: &mut W) -> Result<(), <W as Write>::Err>
    where
        W: Write;
}

The Serial trait provides a means of writing structures into byte-sinks (Write).

Can be derived using #[derive(Serial)] for most cases.

Required methods

fn serial<W>(&self, _out: &mut W) -> Result<(), <W as Write>::Err> where
W: Write,

Attempt to write the structure into the provided writer, failing if only part of the structure could be written.

NB: We use Result instead of Option for better composability with other constructs.