Trait concordium_rust_sdk::types::smart_contracts::concordium_contracts_common::SerialCtx

pub trait SerialCtx {
    fn serial_ctx<W>(
        &self, 
        size_length: SizeLength, 
        out: &mut W
    ) -> Result<(), <W as Write>::Err>
    where
        W: Write;
}

The SerialCtx trait provides a means of writing structures into byte-sinks (Write) using contextual information. The contextual information is:

size_length: The number of bytes used to record the length of the data.

Required methods

fn serial_ctx<W>(
&self,
size_length: SizeLength,
out: &mut W
) -> Result<(), <W as Write>::Err> where
W: Write,

Attempt to write the structure into the provided writer, failing if if the length cannot be represented in the provided size_length or only part of the structure could be written.

NB: We use Result instead of Option for better composability with other constructs.