Module concordium_rust_sdk::types::transactions::cost

Energy costs of transactions.

Constants

A
The B constant for NRG assignment. This scales the effect of the number of signatures on the energy.

ADD_BAKER
Additional cost of registerding the account as a baker.

B
The A constant for NRG assignment. This scales the effect of transaction size on the energy.

CONFIGURE_BAKER_WITHOUT_KEYS
Additional cost of configuring a baker if new keys are not registered.

CONFIGURE_BAKER_WITH_KEYS
Additional cost of configuring a baker if new keys are registered.

CONFIGURE_DELEGATION
Additional cost of configuring delegation.

ENCRYPTED_TRANSFER
Additional cost of an encrypted transfer.

REGISTER_DATA
Additional cost of registering a piece of data.

REMOVE_BAKER
Additional cost of removing a baker.

SIMPLE_TRANSFER
Additional cost of a normal, account to account, transfer.

TRANSFER_TO_ENCRYPTED
Additional cost of a transfer from public to encrypted balance.

TRANSFER_TO_PUBLIC
Additional cost of a transfer from encrypted to public balance.

UPDATE_BAKER_KEYS
Additional cost of updating baker’s keys.

UPDATE_BAKER_RESTAKE
Additional cost of updating the baker’s restake flag.

UPDATE_BAKER_STAKE
Additional cost of updating the baker’s stake, either increasing or lowering it.

Functions

base_cost
Base cost of a transaction is the minimum cost that accounts for transaction size and signature checking. In addition to base cost each transaction has a transaction-type specific cost.

deploy_credential
Additional cost of deploying a credential of the given type and with the given number of keys.

deploy_module
Additional cost of deploying a smart contract module, parametrized by the size of the module, which is defined to be the size of the binary .wasm file that is sent as part of the transaction.

scheduled_transfer
Cost of a scheduled transfer, parametrized by the number of releases.

update_credentials
Additional cost of updating account’s credentials, parametrized by