Enum concordium_rust_sdk::types::transactions::Payload

pub enum Payload {

    DeployModule {
        module: WasmModule,
    },
    InitContract {
        payload: InitContractPayload,
    },
    Update {
        payload: UpdateContractPayload,
    },
    Transfer {
        to_address: AccountAddress,
        amount: Amount,
    },
    AddBaker {
        payload: Box<AddBakerPayload, Global>,
    },
    RemoveBaker,
    UpdateBakerStake {
        stake: Amount,
    },
    UpdateBakerRestakeEarnings {
        restake_earnings: bool,
    },
    UpdateBakerKeys {
        payload: Box<BakerKeysPayload<UpdateBakerKeysMarker>, Global>,
    },
    UpdateCredentialKeys {
        cred_id: CredentialRegistrationID,
        keys: CredentialPublicKeys,
    },
    EncryptedAmountTransfer {
        to: AccountAddress,
        data: Box<EncryptedAmountTransferData<G1>, Global>,
    },
    TransferToEncrypted {
        amount: Amount,
    },
    TransferToPublic {
        data: Box<SecToPubAmountTransferData<G1>, Global>,
    },
    TransferWithSchedule {
        to: AccountAddress,
        schedule: Vec<(Timestamp, Amount), Global>,
    },
    UpdateCredentials {
        new_cred_infos: BTreeMap<CredentialIndex, CredentialDeploymentInfo<Bls12, G1, AttributeKind>>,
        remove_cred_ids: Vec<CredentialRegistrationID, Global>,
        new_threshold: AccountThreshold,
    },
    RegisterData {
        data: RegisteredData,
    },
    TransferWithMemo {
        to_address: AccountAddress,
        memo: Memo,
        amount: Amount,
    },
    EncryptedAmountTransferWithMemo {
        to: AccountAddress,
        memo: Memo,
        data: Box<EncryptedAmountTransferData<G1>, Global>,
    },
    TransferWithScheduleAndMemo {
        to: AccountAddress,
        memo: Memo,
        schedule: Vec<(Timestamp, Amount), Global>,
    },
    ConfigureBaker {
        data: Box<ConfigureBakerPayload, Global>,
    },
    ConfigureDelegation {
        data: ConfigureDelegationPayload,
    },

}

Payload of an account transaction.

Variants

DeployModule

Deploy a Wasm module with the given source.

Fields of DeployModule

module: WasmModule

InitContract

Initialize a new smart contract instance.

Fields of InitContract

payload: InitContractPayload

Update

Update a smart contract instance by invoking a specific function.

Fields of Update

payload: UpdateContractPayload

Transfer

Transfer CCD to an account.

Fields of Transfer

to_address: AccountAddress

Address to send to.

amount: Amount

Amount to send.

AddBaker

Register the sender account as a baker.

Fields of AddBaker

payload: Box<AddBakerPayload, Global>

RemoveBaker

Deregister the account as a baker.

UpdateBakerStake

Update baker’s stake.

Fields of UpdateBakerStake

stake: Amount

The new stake.

UpdateBakerRestakeEarnings

Modify whether to add earnings to the baker stake automatically or not.

Fields of UpdateBakerRestakeEarnings

restake_earnings: bool

New value of the flag.

UpdateBakerKeys

Update the baker’s keys.

Fields of UpdateBakerKeys

payload: Box<BakerKeysPayload<UpdateBakerKeysMarker>, Global>

UpdateCredentialKeys

Update signing keys of a specific credential.

Fields of UpdateCredentialKeys

cred_id: CredentialRegistrationID

Id of the credential whose keys are to be updated.

keys: CredentialPublicKeys

The new public keys.

EncryptedAmountTransfer

Transfer an encrypted amount.

Fields of EncryptedAmountTransfer

to: AccountAddress

The recepient’s address.

data: Box<EncryptedAmountTransferData<G1>, Global>

The (encrypted) amount to transfer and proof of correctness of accounting.

TransferToEncrypted

Transfer from public to encrypted balance of the sender account.

Fields of TransferToEncrypted

amount: Amount

The amount to transfer.

TransferToPublic

Transfer an amount from encrypted to the public balance of the account.

Fields of TransferToPublic

data: Box<SecToPubAmountTransferData<G1>, Global>

The amount to transfer and proof of correctness of accounting.

TransferWithSchedule

Transfer an amount with schedule.

Fields of TransferWithSchedule

to: AccountAddress

The recepient.

schedule: Vec<(Timestamp, Amount), Global>

The release schedule. This can be at most 255 elements.

UpdateCredentials

Update the account’s credentials.

Fields of UpdateCredentials

new_cred_infos: BTreeMap<CredentialIndex, CredentialDeploymentInfo<Bls12, G1, AttributeKind>>

New credentials to add.

remove_cred_ids: Vec<CredentialRegistrationID, Global>

Ids of credentials to remove.

new_threshold: AccountThreshold

The new account threshold.

RegisterData

Register the given data on the chain.

Fields of RegisterData

data: RegisteredData

The data to register.

TransferWithMemo

Transfer CCD to an account with an additional memo.

Fields of TransferWithMemo

to_address: AccountAddress

Address to send to.

memo: Memo

Memo to include in the transfer.

amount: Amount

Amount to send.

EncryptedAmountTransferWithMemo

Transfer an encrypted amount.

Fields of EncryptedAmountTransferWithMemo

to: AccountAddress

The recepient’s address.

memo: Memo

Memo to include in the transfer.

data: Box<EncryptedAmountTransferData<G1>, Global>

The (encrypted) amount to transfer and proof of correctness of accounting.

TransferWithScheduleAndMemo

Transfer an amount with schedule.

Fields of TransferWithScheduleAndMemo

to: AccountAddress

The recepient.

memo: Memo

Memo to include in the transfer.

schedule: Vec<(Timestamp, Amount), Global>

The release schedule. This can be at most 255 elements.

ConfigureBaker

Configure a baker on an account.

Fields of ConfigureBaker

data: Box<ConfigureBakerPayload, Global>

ConfigureDelegation

Configure an account’s stake delegation.

Fields of ConfigureDelegation

data: ConfigureDelegationPayload

Implementations

impl Payload

pub fn transaction_type(&self) -> TransactionType

Resolve the TransactionType corresponding to the variant of the Payload.