Function concordium_rust_sdk::types::transactions::send::update_contract

pub fn update_contract(
    signer: &impl ExactSizeTransactionSigner, 
    sender: AccountAddress, 
    nonce: Nonce, 
    expiry: TransactionTime, 
    payload: UpdateContractPayload, 
    energy: Energy
) -> AccountTransaction<EncodedPayload>

Update a smart contract intance, giving it the given amount of energy for execution. The unique parameters are

energy – the amount of energy that can be used for contract execution. The base energy amount for transaction verification will be added to this cost.