Trait concordium_rust_sdk::types::transactions::ExactSizeTransactionSigner

pub trait ExactSizeTransactionSigner: TransactionSigner {
    fn num_keys(&self) -> u32;
}

A signing implementation that knows the number of keys up-front.

Required methods

fn num_keys(&self) -> u32

Return the number of keys that the signer will sign with. This must match what TransactionSigner::sign_transaction_hash returns.

Loading content...