Struct concordium_rust_sdk::v2::Client

pub struct Client { /* fields omitted */ }

A client for gRPC API v2 of the Concordium node. Can be used to control the node, send transactions and query information about the node and the state of the chain.

Connecting to a Concordium node

Creates a new client connection to a Concordium node. Make sure to have access to the gRPC API v2 endpoint of a running node.

use concordium_rust_sdk::v2::{Client, Endpoint};
use std::str::FromStr;

// Assumes the node is running locally and gRPC API v2 can be accessed on port 20001.
let node_endpoint = Endpoint::from_str("http://localhost:20001")?;
let mut client = Client::new(node_endpoint).await?;

// Verify the connection to the node by printing node information.
let node_info = client.get_node_info().await?;
println!("{:#?}", node_info);

Concurrent use of the client

All endpoints take a &mut self as an argument which means that a single instance cannot be used concurrently. However instead of putting the Client behind a Mutex, the intended way to use it is to clone it. Cloning is very cheap and will reuse the underlying connection.

Implementations

impl Client

pub async fn new<E: Into<Endpoint>>(endpoint: E) -> Result<Self, Error>

Construct a new client connection to a concordium node.

Example

Creates a new client. Note the example assumes access to a local running node.

use concordium_rust_sdk::{endpoints::Endpoint, v2::Client};
use std::str::FromStr;

let node_endpoint = Endpoint::from_str("http://localhost:20001")?;
let mut client = Client::new(node_endpoint).await?;

pub async fn get_account_info(
&mut self,
acc: &AccountIdentifier,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<AccountInfo>>

Get the information for the given account in the given block. If either the block or the account do not exist QueryError::NotFound is returned.

pub async fn get_next_account_sequence_number(
&mut self,
account_address: &AccountAddress
) -> QueryResult<AccountNonceResponse>

Get the next sequence number for the account, with information on how reliable the information is.

pub async fn get_consensus_info(&mut self) -> QueryResult<ConsensusInfo>

Get information about the current state of consensus. This is an overview of the node’s current view of the chain.

pub async fn get_cryptographic_parameters(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<CryptographicParameters>>

Get the currently used cryptographic parameters. If the block does not exist QueryError::NotFound is returned.

pub async fn get_account_list(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<AccountAddress, Status>>>>

Get the list of accounts in the given block. The stream will end when all accounts that exist in the state at the end of the given block have been returned. If the block does not exist QueryError::NotFound is returned.

pub async fn get_module_list(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<ModuleRef, Status>>>>

Get a list of all smart contract modules. The stream will end when all modules that exist in the state at the end of the given block have been returned. If the block does not exist QueryError::NotFound is returned.

pub async fn get_module_source(
&mut self,
module_ref: &ModuleRef,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<WasmModule>>

Get the source of a smart contract module. If the block or module does not exist QueryError::NotFound is returned.

pub async fn get_instance_list(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<ContractAddress, Status>>>>

Get the list of smart contract instances in a given block. The stream will end when all instances that exist in the state at the end of the given block have been returned. If the block does not exist QueryError::NotFound is returned.

pub async fn get_instance_info(
&mut self,
address: ContractAddress,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<InstanceInfo>>

Get information about a smart contract instance as it appears at the end of the given block. If the block or instance does not exist QueryError::NotFound is returned.

pub async fn get_ancestors(
&mut self,
bi: &BlockIdentifier,
limit: u64
) -> QueryResult<QueryResponse<impl Stream<Item = Result<BlockHash, Status>>>>

Get a stream of ancestors for the provided block. Starting with the provided block itself, moving backwards until no more ancestors or the requested number of ancestors have been returned.

pub async fn get_finalized_blocks(
&mut self
) -> QueryResult<impl Stream<Item = Result<FinalizedBlockInfo, Status>>>

Return a stream of blocks that are finalized from the time the query is made onward. This can be used to listen for newly finalized blocks.

Note: There is no guarantee that blocks will not be skipped if the client is too slow in processing the stream, however blocks will always be sent by increasing block height.

pub async fn get_instance_state(
&mut self,
ca: ContractAddress,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<(Vec<u8>, Vec<u8>), Status>>>>

Get the exact state of a specific contract instance, streamed as a list of key-value pairs. The list is streamed in lexicographic order of keys. If the block or instance does not exist QueryError::NotFound is returned.

pub async fn instance_state_lookup(
&mut self,
ca: ContractAddress,
key: impl Into<Vec<u8>>,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<Vec<u8>>>

Get the value at a specific key of a contract state. In contrast to get_instance_state this is more efficient, but requires the user to know the specific key to look for. If the block or instance does not exist QueryError::NotFound is returned.

pub async fn get_block_item_status(
&mut self,
th: &TransactionHash
) -> QueryResult<TransactionStatus>

Get the status of and information about a specific block item (transaction). If the block item does not exist QueryError::NotFound is returned.

pub async fn send_block_item<P: PayloadLike>(
&mut self,
bi: &BlockItem<P>
) -> RPCResult<TransactionHash>

Send a block item. A block item is either an AccountTransaction, which is a transaction signed and paid for by an account, a CredentialDeployment, which creates a new account, or UpdateInstruction, which is an instruction to change some parameters of the chain. Update instructions can only be sent by the governance committee.

pub async fn send_account_transaction<P: PayloadLike>(
&mut self,
at: AccountTransaction<P>
) -> RPCResult<TransactionHash>

Send an account transaction. This is just a helper around send_block_item block item for convenience.

pub async fn get_account_transaction_sign_hash(
&mut self,
header: &TransactionHeader,
payload: &Payload
) -> RPCResult<TransactionSignHash>

Get the hash to be signed for an account transaction from the node. The hash returned can then be used for signing when constructing TransactionSignature as part of calling Client::send_block_item.

This is provided as a convenience to support cases where the right SDK is not available for interacting with the node.

This SDK can compute the hash off-line and it is not recommended to use this endpoint, instead use compute_transaction_sign_hash.

pub async fn wait_until_finalized(
&mut self,
hash: &TransactionHash
) -> QueryResult<(BlockHash, BlockItemSummary)>

Wait until the transaction is finalized. Returns NotFound in case the transaction is not known to the node. In case of success, the return value is a pair of the block hash of the block that contains the transactions, and its outcome in the block.

Since this can take an indefinite amount of time in general, users of this function might wish to wrap it inside timeout handler and handle the resulting failure.

pub async fn invoke_instance(
&mut self,
bi: &BlockIdentifier,
context: &ContractContext
) -> QueryResult<QueryResponse<InvokeContractResult>>

Run the smart contract instance entrypoint in a given context and in the state at the end of the given block and return the results. If the block does not exist QueryError::NotFound is returned.

pub async fn get_block_info(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<BlockInfo>>

Get information, such as height, timings, and transaction counts for the given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_baker_list(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<BakerId, Status>>>>

Get all the bakers at the end of the given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_pool_info(
&mut self,
block_id: &BlockIdentifier,
baker_id: BakerId
) -> QueryResult<QueryResponse<BakerPoolStatus>>

Get information about a given pool at the end of a given block. If the block or baker ID does not exist QueryError::NotFound is returned.

pub async fn get_passive_delegation_info(
&mut self,
block_id: &BlockIdentifier
) -> QueryResult<QueryResponse<PassiveDelegationStatus>>

Get information about the passive delegators at the end of a given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_blocks_at_height(
&mut self,
blocks_at_height_input: &BlocksAtHeightInput
) -> QueryResult<Vec<BlockHash>>

Get a list of live blocks at a given height.

pub async fn get_tokenomics_info(
&mut self,
block_id: &BlockIdentifier
) -> QueryResult<QueryResponse<RewardsOverview>>

Get information about tokenomics at the end of a given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_pool_delegators(
&mut self,
bi: &BlockIdentifier,
baker_id: BakerId
) -> QueryResult<QueryResponse<impl Stream<Item = Result<DelegatorInfo, Status>>>>

Get the registered delegators of a given pool at the end of a given block. If the block or baker ID does not exist QueryError::NotFound is returned. The stream will end when all the delegators have been returned for the given block.

In contrast to the Client::get_pool_delegators_reward_period which returns delegators that are fixed for the reward period of the block, this endpoint returns the list of delegators that are registered in the block. Any changes to delegators are immediately visible in this list.

pub async fn get_pool_delegators_reward_period(
&mut self,
bi: &BlockIdentifier,
baker_id: BakerId
) -> QueryResult<QueryResponse<impl Stream<Item = Result<DelegatorRewardPeriodInfo, Status>>>>

Get the fixed delegators of a given pool for the reward period of the given block. If the block or baker ID does not exist QueryError::NotFound is returned. The stream will end when all the delegators have been returned.

In contracts to the Client::get_pool_delegators which returns delegators registered for the given block, this endpoint returns the fixed delegators contributing stake in the reward period containing the given block.

pub async fn get_passive_delegators(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<DelegatorInfo, Status>>>>

Get the registered passive delegators at the end of a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the delegators have been returned.

In contrast to the Client::get_passive_delegators_reward_period which returns delegators that are fixed for the reward period of the block, this endpoint returns the list of delegators that are registered in the block. Any changes to delegators are immediately visible in this list.

pub async fn get_passive_delegators_reward_period(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<DelegatorRewardPeriodInfo, Status>>>>

Get the fixed passive delegators for the reward period of the given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the delegators have been returned.

In contracts to the GetPassiveDelegators which returns delegators registered for the given block, this endpoint returns the fixed delegators contributing stake in the reward period containing the given block.

pub async fn get_branches(&mut self) -> QueryResult<Branch>

Get the current branches of blocks starting from and including the last finalized block.

Branches are all live blocks that are successors of the last finalized block. In particular this means that blocks which do not have a parent are not included in this response.

pub async fn get_election_info(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<BirkParameters>>

Get information related to the baker election for a particular block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_identity_providers(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<IpInfo<IpPairing>, Status>>>>

Get the identity providers registered as of the end of a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the identity providers have been returned.

pub async fn get_anonymity_revokers(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<ArInfo<ArCurve>, Status>>>>

Get the list of anonymity revokers in the given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the anonymity revokers have been returned.

pub async fn get_account_non_finalized_transactions(
&mut self,
account_address: &AccountAddress
) -> QueryResult<impl Stream<Item = Result<TransactionHash, Status>>>

Get the list of transactions hashes for transactions that claim to be from the given account, but which are not yet finalized. They are either committed to a block or still pending. The stream will end when all the non-finalized transaction hashes have been returned. If the account does not exist an empty list will be returned.

This endpoint is not expected to return a large amount of data in most cases, but in bad network condtions it might.

pub async fn get_block_items(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<BlockItem<EncodedPayload>, Status>>>>

Get the block items included in a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the block items in the given block have been returned.

pub async fn shutdown(&mut self) -> RPCResult<()>

Shut down the node. Return a GRPC error if the shutdown failed.

pub async fn peer_connect(&mut self, addr: SocketAddr) -> RPCResult<()>

Suggest a peer to connect to the submitted peer details. This, if successful, adds the peer to the list of given addresses. Otherwise return a GRPC error.

Note: The peer might not be connected to instantly, in that case the node will try to establish the connection in near future. This function returns a GRPC status ‘Ok’ in this case.

pub async fn peer_disconnect(&mut self, addr: SocketAddr) -> RPCResult<()>

Disconnect from the peer and remove them from the given addresses list if they are on it. Return if the request was processed successfully. Otherwise return a GRPC error.

pub async fn get_banned_peers(&mut self) -> RPCResult<Vec<BannedPeer>>

Get a vector of the banned peers.

pub async fn ban_peer(&mut self, peer_to_ban: PeerToBan) -> RPCResult<()>

Ban a peer. When successful return Ok(()), and otherwise return an error describing the issue.

pub async fn unban_peer(&mut self, banned_peer: &BannedPeer) -> RPCResult<()>

Unban a peer. When successful return Ok(()), and otherwise return an error describing the issue.

pub async fn dump_start(&mut self, file: &Path, raw: bool) -> RPCResult<()>

Start a network dump if the feature is enabled on the node. This writes all the network packets into the given file. Return Ok(()) if a network dump has been initiated, and an error otherwise.

file - The file to write to.

raw - Whether raw packets should be included in the dump or not.

Note. If the feature ‘network_dump’ is not enabled on the node then this will return a ‘Precondition failed’ error.

pub async fn dump_stop(&mut self) -> RPCResult<()>

Stop an ongoing network dump. Return nothing if it was successfully stopped, and otherwise return an error.

Note. If the feature ‘network_dump’ is not enabled on the node then this will return a ‘Precondition failed’ error.

pub async fn get_peers_info(&mut self) -> RPCResult<PeersInfo>

Get a list of the peers that the node is connected to and associated network related information for each peer.

pub async fn get_node_info(&mut self) -> RPCResult<NodeInfo>

Retrieve information about the node. The response contains meta information about the node such as the version of the software, the local time of the node etc.

The response also yields network related information such as the node ID, bytes sent/received etc.

Finally depending on the type of the node (regular node or ‘bootstrapper’) the response also yields baking information if the node is configured with baker credentials.

Bootstrappers do no reveal any consensus information as they do not run the consensus protocol.

pub async fn get_block_transaction_events(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<BlockItemSummary, Status>>>>

Get the transaction events in a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the transaction events for a given block have been returned.

pub async fn get_block_special_events(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<SpecialTransactionOutcome, Status>>>>

Get a the special events in a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the special events for a given block have been returned.

These are events generated by the protocol, such as minting and reward payouts. They are not directly generated by any transaction.

pub async fn get_block_pending_updates(
&mut self,
bi: &BlockIdentifier
) -> QueryResult<QueryResponse<impl Stream<Item = Result<PendingUpdate, Status>>>>

Get the pending updates to chain parameters at the end of a given block. If the block does not exist QueryError::NotFound is returned. The stream will end when all the pending updates for a given block have been returned.

pub async fn get_next_update_sequence_numbers(
&mut self,
block_id: &BlockIdentifier
) -> QueryResult<QueryResponse<NextUpdateSequenceNumbers>>

Get next available sequence numbers for updating chain parameters after a given block. If the block does not exist then QueryError::NotFound is returned.

pub async fn get_block_chain_parameters(
&mut self,
block_id: &BlockIdentifier
) -> QueryResult<QueryResponse<ChainParameters>>

Get the chain parameters in effect after a given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_block_finalization_summary(
&mut self,
block_id: &BlockIdentifier
) -> QueryResult<QueryResponse<Option<FinalizationSummary>>>

Get the information about a finalization record in a block. A block can contain zero or one finalization record. If a record is contained then this query will return information about the finalization session that produced it, including the finalizers eligible for the session, their power, and whether they signed this particular record. If the block does not exist QueryError::NotFound is returned.

pub async fn get_finalized_blocks_from(
&mut self,
start_height: AbsoluteBlockHeight
) -> QueryResult<FinalizedBlocksStream>

Get a continous stream of finalized blocks starting from a given height. This function starts a background task (a tokio task) that listens for new finalized blocks. This task is killed when the FinalizedBlocksStream is dropped.