Module concordium_rust_sdk::endpoints

Wrapper for the node’s GRPC API. The return values are parsed and wrapped in structured values.

Structs

Client

Client that can perform queries. All endpoints take a &mut self as an argument which means that a single instance cannot be used concurrently. However instead of putting the Client behind a Mutex, the intended way to use it is to clone it. Cloning is very cheap and will reuse the underlying connection.

Endpoint

Channel builder.

Enums

BlocksAtHeightInput

Input to the Client::get_blocks_at_height query.

QueryError

Errors that can occur when making queries. This can either be a general connection/authentication error, or the requested item is not found.

RPCError

Authentication, connection, or response parsing error.

Type Definitions

QueryResult

Result a GRPC query where the item lookup might fail. This is a simple alias for std::Result that fixes the error type to be QueryError.

RPCResult

Result a GRPC query. This is a simple alias for std::Result that fixes the error type to be RPCError.