Struct concordium_rust_sdk::endpoints::Client

pub struct Client { /* fields omitted */ }

Client that can perform queries. All endpoints take a &mut self as an argument which means that a single instance cannot be used concurrently. However instead of putting the Client behind a Mutex, the intended way to use it is to clone it. Cloning is very cheap and will reuse the underlying connection.

Implementations

impl Client

pub async fn connect<D: TryInto<Endpoint>>(
dst: D,
token: impl Into<String>
) -> Result<Self, Error> where
<D as TryInto<Endpoint>>::Error: Error + Send + Sync + 'static,

Construct a new client by connecting to the specified destination. The provided token must match the authentication token accepted by the node.

pub async fn peer_connect(&mut self, ip: &IpAddr, port: u16) -> RPCResult<bool>

Instruct the node to try to connect to the given peer. This also adds the address to the list of trusted addresses. These are addresses to which the node will try to keep connected to at all times.

pub async fn uptime(&mut self) -> RPCResult<Duration>

Query for the node’s uptime.

pub async fn total_sent(&mut self) -> RPCResult<u64>

Query for the total number of packets that the node has sent thus far.

pub async fn total_received(&mut self) -> RPCResult<u64>

Query for the total number of packets that the node has received thus far.

pub async fn version(&mut self) -> RPCResult<Version>

Query for the node version.

pub async fn peer_statistics(
&mut self,
include_bootstrappers: bool
) -> RPCResult<PeerStatsResponse>

FIXME: Better return type.

pub async fn peer_list(
&mut self,
include_bootstrappers: bool
) -> RPCResult<Vec<PeerElement>>

Get the list of peers, possibly including any bootstrappers the node is currently connected to.

pub async fn ban_node(&mut self, ban_method: BanMethod) -> RPCResult<bool>

Ban a specific node. See Client::unban_node for the dual. Note that this will also cause the node to drop any connections to a matching node.

pub async fn unban_node(&mut self, ip: IpAddr) -> RPCResult<bool>

Unban a specific node. See Client::ban_node for the dual. Note that this will only remove the ban, but it will not re-establish any connections. To connect use Client::peer_connect.

pub async fn join_network(&mut self, network_id: NetworkId) -> RPCResult<bool>

Ask the node to join the specified network.

pub async fn leave_network(&mut self, network_id: NetworkId) -> RPCResult<bool>

Ask the node to leave the specified network.

pub async fn node_info(&mut self) -> RPCResult<NodeInfo>

Get some general information about a running node. See the return type for details of the data that is returned.

pub async fn get_consensus_status(&mut self) -> RPCResult<ConsensusInfo>

Get consensus information from the node. This is an overview of the node’s view of the chain.

pub async fn get_block_info(
&mut self,
block_hash: &BlockHash
) -> QueryResult<BlockInfo>

Get information about a specific block, if it exists.

pub async fn get_ancestors(
&mut self,
block: &BlockHash,
num: u64
) -> QueryResult<Vec<BlockHash>>

Get the ancestors of a given block, if any.

pub async fn get_branches(&mut self) -> RPCResult<Branch>

Get the branches of the node’s tree. Branches are all live blocks that are successors of the last finalized block. In particular this means that blocks which do not have a parent are not included in this response.

pub async fn get_blocks_at_height(
&mut self,
bh: BlocksAtHeightInput
) -> RPCResult<Vec<BlockHash>>

Get the list of block hashes at the given height. If there are no blocks at the given height an empty list is returned.

pub async fn stop_baker(&mut self) -> RPCResult<bool>

Stop the baker thread. The node will still keep running and responding to queries.

pub async fn get_banned_ips(&mut self) -> RPCResult<Vec<IpAddr>>

Get the list of ips that the node will currently not connect to, nor accept connections for. See Client::ban_node and Client::unban_node for functions to add and remove ips from the list.

pub async fn shutdown(&mut self) -> RPCResult<bool>

Stop the node. After this is called the node will stop.

pub async fn get_account_list(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<AccountAddress>>

Get the list of accounts in the given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_instances(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<ContractAddress>>

Get the list of smart contract instances in a given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_account_info(
&mut self,
addr: impl Borrow<AccountAddress>,
bh: &BlockHash
) -> QueryResult<AccountInfo>

Get the information for the given account in the given block. If either the block or the account does not exist QueryError::NotFound is returned.

pub async fn get_account_info_raw(
&mut self,
addr: impl Borrow<AccountAddress>,
bh: &BlockHash
) -> QueryResult<Value>

Get the information for the given account in the given block. If either the block or the account does not exist QueryError::NotFound is returned. In contrast to Client::get_account_info this function does not fully parse the result. Since parsing account responses can be relatively expensive this provides an option to delay parsing, or to only take out individual values. The return Value will always be parseable as a AccountInfo

pub async fn get_account_info_by_cred_id(
&mut self,
addr: impl Borrow<CredentialRegistrationID>,
bh: &BlockHash
) -> QueryResult<AccountInfo>

Get the information for the given account in the given block by credential registration id. This returns an account that is either currently, or was in the past, associated with a credential with the specified registration id.

If either the block or the credential registration id does no not exist then QueryError::NotFound is returned.

pub async fn get_account_info_by_cred_id_raw(
&mut self,
addr: impl Borrow<CredentialRegistrationID>,
bh: &BlockHash
) -> QueryResult<Value>

Get the information for the given account in the given block by credential registration id. This returns an account that is either currently, or was in the past, associated with a credential with the specified registration id.

If either the block or the credential registration id does no not exist then QueryError::NotFound is returned.

Compares to Client::get_account_info_by_cred_id analogously to how Client::get_account_info compares to Client::get_account_info_raw

pub async fn get_instance_info(
&mut self,
addr: ContractAddress,
bh: &BlockHash
) -> QueryResult<InstanceInfo>

Get the information for the given smart contract instance in the given block.

pub async fn get_reward_status(
&mut self,
bh: &BlockHash
) -> QueryResult<RewardsOverview>

Get the information about total amount of CCD and the state of various administrative accounts.

pub async fn get_birk_parameters(
&mut self,
bh: &BlockHash
) -> QueryResult<BirkParameters>

Get consensus-relevant information for the specified block.

pub async fn get_baker_list(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<BakerId>>

Get the IDs of the bakers registered in the given block. Note that this list is in general different from the bakers that are returned as part of get_birk_parameters. The latter are the bakers that are eligible for baking at the time of the block.

pub async fn get_pool_status(
&mut self,
baker_id: Option<BakerId>,
bh: &BlockHash
) -> QueryResult<PoolStatus>

Get the status of a given baker pool or passive delegation at the given block. This is only supported if the node’s protocol version is at least P4. For earlier versions it will return NotFound.

pub async fn get_module_list(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<ModuleRef>>

Get the list of smart contract modules in the given block.

pub async fn get_module_source(
&mut self,
mr: &ModuleRef,
bh: &BlockHash
) -> QueryResult<Vec<u8>>

pub async fn get_identity_providers(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<IpInfo<IpPairing>>>

Get the list of identity providers in the given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_anonymity_revokers(
&mut self,
bh: &BlockHash
) -> QueryResult<Vec<ArInfo<ArCurve>>>

Get the list of anonymity revokers in the given block. If the block does not exist QueryError::NotFound is returned.

pub async fn get_cryptographic_parameters(
&mut self,
bh: &BlockHash
) -> QueryResult<GlobalContext<ArCurve>>

Get the currently used cryptographic parameters. If the block does not exist QueryError::NotFound is returned.

pub async fn get_account_non_finalized_transactions(
&mut self,
addr: &AccountAddress
) -> RPCResult<Vec<TransactionHash>>

Get the list of transactions hashes for transactions that claim to be from the given account, but which are not yet finalized. They are either committed to a block or still pending.

If the account does not exist an empty list will be returned.

pub async fn get_transaction_status_in_block(
&mut self,
bh: &BlockHash,
th: &TransactionHash
) -> QueryResult<TransactionStatusInBlock>

Get the status of a transaction in a given block. If the transaction is not known, or is committed or finalized, but not in the given block, QueryError::NotFound is returned.

pub async fn get_transaction_status(
&mut self,
th: &TransactionHash
) -> QueryResult<TransactionStatus>

Query the status of the transaction. If the transaction is not known to the node QueryError::NotFound is returned.

pub async fn get_block_summary(
&mut self,
bh: &BlockHash
) -> QueryResult<BlockSummary>

Get the summary of a block. This lists all transactions and special outcomes occurring in a given block, as well as the value of chain parameters at the given block. If the block is not in the node’s tree QueryError::NotFound is returned.

pub async fn get_block_summary_raw(
&mut self,
bh: &BlockHash
) -> QueryResult<Value>

Get the summary of a block. This lists all transactions and special outcomes occurring in a given block, as well as the value of chain parameters at the given block. If the block is not in the node’s tree QueryError::NotFound is returned. In contrast to Client::get_block_summary this function does not fully parse the result since that can be expensive and some applications might not need it.

pub async fn get_next_account_nonce(
&mut self,
addr: &AccountAddress
) -> RPCResult<AccountNonceResponse>

Get the next nonce for the account, with information on how reliable the information is.

pub async fn invoke_contract(
&mut self,
bh: &BlockHash,
context: &ContractContext
) -> RPCResult<InvokeContractResult>

Invoke a contract in the given context and get the response. Note that this is purely node-local transient operation and no transaction or payment is involved. Use send_transaction if you need to update the state of the contract on the chain.

pub async fn send_block_item<PayloadType: PayloadLike>(
&mut self,
bi: &BlockItem<PayloadType>
) -> RPCResult<TransactionHash>

Send the given block item.

pub async fn send_raw_account_transaction(
&mut self,
signatures: &TransactionSignature,
body: &[u8]
) -> RPCResult<TransactionHash>

Send the given account transaction item on the given network. This is a low-level function that can be useful in case a transaction is constructed by a third party. It avoids deserializing and converting data. If the transaciton is accepted by the node then the transaction hash that can be used to query the status is returned.