Enum concordium_rust_sdk::types::AccountTransactionEffects

pub enum AccountTransactionEffects {

    None {
        transaction_type: Option<TransactionType>,
        reject_reason: RejectReason,
    },
    ModuleDeployed {
        module_ref: ModuleRef,
    },
    ContractInitialized {
        data: ContractInitializedEvent,
    },
    ContractUpdateIssued {
        effects: Vec<ContractTraceElement>,
    },
    AccountTransfer {
        amount: Amount,
        to: AccountAddress,
    },
    AccountTransferWithMemo {
        amount: Amount,
        to: AccountAddress,
        memo: Memo,
    },
    BakerAdded {
        data: Box<BakerAddedEvent>,
    },
    BakerRemoved {
        baker_id: BakerId,
    },
    BakerStakeUpdated {
        data: Option<BakerStakeUpdatedData>,
    },
    BakerRestakeEarningsUpdated {
        baker_id: BakerId,
        restake_earnings: bool,
    },
    BakerKeysUpdated {
        data: Box<BakerKeysEvent>,
    },
    EncryptedAmountTransferred {
        removed: Box<EncryptedAmountRemovedEvent>,
        added: Box<NewEncryptedAmountEvent>,
    },
    EncryptedAmountTransferredWithMemo {
        removed: Box<EncryptedAmountRemovedEvent>,
        added: Box<NewEncryptedAmountEvent>,
        memo: Memo,
    },
    TransferredToEncrypted {
        data: Box<EncryptedSelfAmountAddedEvent>,
    },
    TransferredToPublic {
        removed: Box<EncryptedAmountRemovedEvent>,
        amount: Amount,
    },
    TransferredWithSchedule {
        to: AccountAddress,
        amount: Vec<(Timestamp, Amount)>,
    },
    TransferredWithScheduleAndMemo {
        to: AccountAddress,
        amount: Vec<(Timestamp, Amount)>,
        memo: Memo,
    },
    CredentialKeysUpdated {
        cred_id: CredentialRegistrationID,
    },
    CredentialsUpdated {
        new_cred_ids: Vec<CredentialRegistrationID>,
        removed_cred_ids: Vec<CredentialRegistrationID>,
        new_threshold: AccountThreshold,
    },
    DataRegistered {
        data: RegisteredData,
    },
    BakerConfigured {
        data: Vec<BakerEvent>,
    },
    DelegationConfigured {
        data: Vec<DelegationEvent>,
    },

}

Effects of an account transactions. All variants apart from AccountTransactionEffects::None correspond to a unique transaction that was successful.

Variants

None

No effects other than payment from this transaction. The rejection reason indicates why the transaction failed.

Fields of None

transaction_type: Option<TransactionType>

Transaction type of a failed transaction, if known. In case of serialization failure this will be None.

reject_reason: RejectReason

Reason for rejection of the transaction

ModuleDeployed

A module was deployed. This corresponds to DeployModule transaction type.

Fields of ModuleDeployed

module_ref: ModuleRef

ContractInitialized

A contract was initialized was deployed. This corresponds to InitContract transaction type.

Fields of ContractInitialized

data: ContractInitializedEvent

ContractUpdateIssued

A contract update transaction was issued and produced the given trace. This is the result of Update transaction.

Fields of ContractUpdateIssued

effects: Vec<ContractTraceElement>

AccountTransfer

A simple account to account transfer occurred. This is the result of a successful Transfer transaction.

Fields of AccountTransfer

amount: Amount

Amount that was transferred.

to: AccountAddress

Receiver account.

AccountTransferWithMemo

A simple account to account transfer occurred with a memo. This is the result of a successful TransferWithMemo transaction.

Fields of AccountTransferWithMemo

amount: Amount

Amount that was transferred.

to: AccountAddress

Receiver account.

memo: Memo

Included memo.

BakerAdded

An account was registered as a baker. This is the result of a successful AddBaker transaction.

Fields of BakerAdded

data: Box<BakerAddedEvent>

BakerRemoved

An account was deregistered as a baker. This is the result of a successful RemoveBaker transaction.

Fields of BakerRemoved

baker_id: BakerId

BakerStakeUpdated

An account was deregistered as a baker. This is the result of a successful UpdateBakerStake transaction.

Fields of BakerStakeUpdated

data: Option<BakerStakeUpdatedData>

If the stake was updated (that is, it changed and did not stay the same) then this is Some, otherwise None.

BakerRestakeEarningsUpdated

An account changed its preference for restaking earnings. This is the result of a successful UpdateBakerRestakeEarnings transaction.

Fields of BakerRestakeEarningsUpdated

baker_id: BakerIdrestake_earnings: bool

The new value of the flag.

BakerKeysUpdated

The baker’s keys were updated. This is the result of a successful UpdateBakerKeys transaction.

Fields of BakerKeysUpdated

data: Box<BakerKeysEvent>

EncryptedAmountTransferred

An encrypted amount was transferred. This is the result of a successful EncryptedAmountTransfer transaction.

Fields of EncryptedAmountTransferred

removed: Box<EncryptedAmountRemovedEvent>added: Box<NewEncryptedAmountEvent>

EncryptedAmountTransferredWithMemo

An encrypted amount was transferred with an included memo. This is the result of a successful EncryptedAmountTransferWithMemo transaction.

Fields of EncryptedAmountTransferredWithMemo

removed: Box<EncryptedAmountRemovedEvent>added: Box<NewEncryptedAmountEvent>memo: Memo

TransferredToEncrypted

An account transferred part of its public balance to its encrypted balance. This is the result of a successful TransferToEncrypted transaction.

Fields of TransferredToEncrypted

data: Box<EncryptedSelfAmountAddedEvent>

TransferredToPublic

An account transferred part of its encrypted balance to its public balance. This is the result of a successful TransferToPublic transaction.

Fields of TransferredToPublic

removed: Box<EncryptedAmountRemovedEvent>amount: Amount

TransferredWithSchedule

A transfer with schedule was performed. This is the result of a successful TransferWithSchedule transaction.

Fields of TransferredWithSchedule

to: AccountAddress

Receiver account.

amount: Vec<(Timestamp, Amount)>

The list of releases. Ordered by increasing timestamp.

TransferredWithScheduleAndMemo

A transfer with schedule was performed with an added memo. This is the result of a successful TransferWithScheduleAndMemo transaction.

Fields of TransferredWithScheduleAndMemo

to: AccountAddress

Receiver account.

amount: Vec<(Timestamp, Amount)>

The list of releases. Ordered by increasing timestamp.

memo: Memo

CredentialKeysUpdated

Keys of a specific credential were updated. This is the result of a successful UpdateCredentialKeys transaction.

Fields of CredentialKeysUpdated

cred_id: CredentialRegistrationID

ID of the credential whose keys were updated.

CredentialsUpdated

Account’s credentials were updated. This is the result of a successful UpdateCredentials transaction.

Fields of CredentialsUpdated

new_cred_ids: Vec<CredentialRegistrationID>

The credential ids that were added.

removed_cred_ids: Vec<CredentialRegistrationID>

The credentials that were removed.

new_threshold: AccountThreshold

The (possibly) updated account threshold.

DataRegistered

Some data was registered on the chain. This is the result of a successful RegisterData transaction.

Fields of DataRegistered

data: RegisteredData

BakerConfigured

A baker was configured. The details of what happened are contained in the list of baker events.

Fields of BakerConfigured

data: Vec<BakerEvent>

DelegationConfigured

An account configured delegation. The details of what happened are contained in the list of delegation events.

Fields of DelegationConfigured

data: Vec<DelegationEvent>

Implementations

impl AccountTransactionEffects

pub fn transaction_type(&self) -> Option<TransactionType>

Get the transaction type corresponding to the effects. Returns None for the AccountTransactionEffects::None variant in case the transaction failed with serialization failure reason.

impl AccountTransactionEffects

pub fn is_rejected(&self) -> Option<&RejectReason>

Return Some if the transaction has been rejected.