Module concordium_rust_sdk::types::queries

Types that appear in various queries of the node.

Structs

AccountNonceResponse

Best guess about the current account nonce, with information about reliability.

BlockInfo

Metadata about a given block.

Branch

Branches of the tree. This is the part of the tree above the last finalized block.

ConsensusInfo

Summary of the current state of consensus.

NextUpdateSequenceNumbers

NodeInfo

Brief information about the node.

PendingUpdate

A scheduled pending update.

Enums

ActiveConsensusState

State of the running node’s consensus when running with baker credentials.

BanMethod

Ways to ban a node, either by IP address or by a peer id.

ConsensusState

Configuration of the node’s consensus.

PeerDetails

Details about the node kind.

PendingUpdateEffect