Module concordium_rust_sdk::types::transactions

Definition of transactions and other transaction-like messages, together with their serialization, signing, and similar auxiliary methods.

Modules

construct
High level wrappers for making transactions with minimal user input. These wrappers handle encoding, setting energy costs when those are fixed for transaction. See also the send module above which combines construction with signing.

cost
Energy costs of transactions.

send
High level wrappers for making transactions with minimal user input. These wrappers handle encoding, setting energy costs when those are fixed for transaction.

update

Structs

AccountAccessStructure
The most straighforward account access structure is a map of public keys with the account threshold.

AccountTransaction
An account transaction signed and paid for by a sender account. The payload type is a generic parameter to support two kinds of payloads, a fully deserialized Payload type, and an EncodedPayload. The latter is useful since deserialization of some types of payloads is expensive. It is thus useful to delay deserialization until after we have checked signatures and the sender account information.

AddBakerPayload
Payload of the AddBaker transaction. This transaction registers the account as a baker.

BakerKeysPayload
Auxiliary type that contains public keys and proof of ownership of those keys. This is used in the AddBaker and UpdateBakerKeys transaction types. The proofs are either constructed for AddBaker or UpdateBakerKeys and the generic V is used as a marker to distinguish this in the type. See the markers: AddBakerKeysMarker and UpdateBakerKeysMarker.

EncodedPayload
An account transaction payload that has not yet been deserialized. This is a simple wrapper around Vec with bespoke serialization.

InitContractPayload
Data needed to initialize a smart contract.

PayloadSize
Type safe wrapper to record the size of the transaction payload.

TransactionHeader
Header of an account transaction that contains basic data to check whether the sender and the transaction is valid.

UpdateContractPayload
Data needed to update a smart contract instance.

UpdateHeader

UpdateInstruction

UpdateInstructionSignature

Enums

AddBakerKeysMarker
Marker for BakerKeysPayload indicating the proofs contained in BakerKeysPayload have been generated for an AddBaker transaction.

BlockItem
A block item are data items that are transmitted on the network either as separate messages, or as part of blocks. They are the only user-generated (as opposed to protocol-generated) message.

Payload
Payload of an account transaction.

UpdateBakerKeysMarker
Marker for BakerKeysPayload indicating the proofs contained in BakerKeysPayload have been generated for an UpdateBakerKeys transaction.

Traits

ExactSizeTransactionSigner
A signing implementation that knows the number of keys up-front.

HasAccountAccessStructure
Implementations of this trait are structures which can produce public keys with which transaction signatures can be verified.

PayloadLike
A helper trait so that we can treat payload and encoded payload in the same place.

TransactionSigner
Abstraction of private keys.

UpdateSigner

Functions

compute_transaction_sign_hash
Compute the transaction sign hash from an encoded payload and header.

get_encoded_payload
Parse an encoded payload of specified length.

sign_transaction
Sign the header and payload, construct the transaction, and return it.

verify_signature_transaction_sign_hash
Verify a signature on the transaction sign hash. This is a low-level operation that is useful to avoid recomputing the transaction hash.

Type Definitions

BakerAddKeysPayload
Baker keys payload containing proofs construct for a AddBaker transaction.

BakerUpdateKeysPayload
Baker keys payload containing proofs construct for a UpdateBakerKeys transaction.